fix: resolve circular dependency
